Conform with trivial golint recommendations.
This is either missing docs or not conforming to the variable naming
suggestions layed out in Effective Go. I moved the package documentation
to a "doc.go", which is pattern I've seen in a bunch of other apps, and
that I think was suggested in some official blog post or another.