Remove UpdateService from PanamaxAdapter.
It's not actually used in the associated server call, so I feel like it
can only lead to sadness.

Scenario A: You spend the time implementing it in your adapter, then
discover that it isn't used.

Scenario B: You implement a dummy version and it gets implemented later
and you don't have a good way of knowing that you now need to implement
it.

Leaving it out of the interface now will avoid the problem. If we need
it later, we add it back to the interface. Then when people upgrade
their code they'll get a clear error that they need to implement
UpdateService.