Allow, but don't require, pmxadapter.Error types.
This will allow us to accept any old regular Go error that the
interfaces want to return, and allow them to just send it through
without having to determine an HTTP code around their errors. We handle
all of that, but can still check to see if a pmxadapter.Error instance
is returned for the more specific error messaging that it can provide.